Authorize to only use location services when active
The ability to use location services when backgrounded is unneeded. This
is a breaking change! You'll need to rename the key in your Info.plist,
as shown here.

> Requesting “Always” authorization is discouraged because of the
potential negative impacts to user privacy. You should request this
level of authorization only when doing so offers a genuine benefit to
the user.